// apps/api/src/generated/endpoints/expenses/expenses.context.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Splitty API
 * OpenAPI spec version: 0.0.0
 */
import type { Context, Env } from 'hono';

import type {
CreateExpenseRequest,
  ExpensesListParams,
  UpdateExpenseRequest
} from '../../schemas';

export type ExpensesListContext<E extends Env = any> = Context<E, '/api/groups/:groupId/expenses', { in: { param: {
 groupId: string,
 },query: ExpensesListParams, }, out: { param: {
 groupId: string,
 },query: ExpensesListParams, } }>
export type ExpensesCreateContext<E extends Env = any> = Context<E, '/api/groups/:groupId/expenses', { in: { param: {
 groupId: string,
 },json: CreateExpenseRequest, }, out: { param: {
 groupId: string,
 },json: CreateExpenseRequest, } }>
export type ExpensesGetContext<E extends Env = any> = Context<E, '/api/groups/:groupId/expenses/:expenseId', { in: { param: {
 groupId: string,
    expenseId: string,
 }, }, out: { param: {
 groupId: string,
    expenseId: string,
 }, } }>
export type ExpensesUpdateContext<E extends Env = any> = Context<E, '/api/groups/:groupId/expenses/:expenseId', { in: { param: {
 groupId: string,
    expenseId: string,
 },json: UpdateExpenseRequest, }, out: { param: {
 groupId: string,
    expenseId: string,
 },json: UpdateExpenseRequest, } }>
export type ExpensesDeleteContext<E extends Env = any> = Context<E, '/api/groups/:groupId/expenses/:expenseId', { in: { param: {
 groupId: string,
    expenseId: string,
 }, }, out: { param: {
 groupId: string,
    expenseId: string,
 }, } }>